requirements.txt
fastapi
mangum
requests
cachetools
pydantic-settings
pytest
redis